Render markdown in the full-message dialog and fix stuck text selection
